requirements.txt
streamlit
google-generativeai
python-dotenv
langchain   # it has lot of functionality to read the pdf
PyPDF2     # to read pdf
chromadb
faiss-cpu    # small no. of file -- gpu for large files
langchain_google_genai  ##to access google-genai api key

# pip install -r requirements.txt
